+++ src/test-data/stdout-3.0.7.txt
@@ -1,0 +1,51 @@
+
+> usfm-tools@1.0.0 generate-markers-map
+> tsx src/generate-markers-map --schema src/test-data/usx-3.0.7.rng --version 3.0.7 --commit 6c490bb5675d281b0fa01876fe67f6e3fd50a4ce --outJSON src/test-data/markers-3.0.7.json
+
+Warning: Marker name "w" has one definition with a default attribute and one without. Using the one with the default attribute.
+Warning: Marker name "rb" has one definition with a default attribute and one without. Using the one with the default attribute.
+Warning: Marker name "ts?(\-[se])?" has one definition with a default attribute and one without. Using the one with the default attribute.
+Warning: Marker name "[Zz].+" has one definition with a default attribute and one without. Using the one with the default attribute.
+Warning: Marker name "qt[1-5]?(\-[se])?" has one definition with a default attribute and one without. Using the one with the default attribute.
+Warning: Style attribute in definition "Figure" has no value or param pattern elements. Skipping.
+Warning: Marker name "ref" has one definition with a default attribute and one without. Using the one with the default attribute.
+Warning: Setting default attribute for jmp to link-href because defaultAttribute was not set
+Warning: Setting default attribute for jmp to link-href because defaultAttribute was not set
+Generated markers.json successfully
+
+Skipped definitions:
+BookHeaders.para.style.enum
+BookIdentification.book.code.enum
+BookIntroduction.para.style.enum
+BookIntroductionEndTitles.para.style.enum
+BookTitles.para.style.enum
+Chapter
+ChapterContent
+ChapterEnd
+Char.char.style.enum
+CharWithAttrib.enum
+CrossReferenceChar.char.style.enum
+Figure
+FootnoteChar.char.style.enum
+IntroChar.char.style.enum
+List.para.style.enum
+ListChar.char.style.enum
+Milestone.ms.style.enum
+MilestoneWithAttrib.enum
+Para.para.style.enum
+Peripheral.BAK.periph.id.enum
+Peripheral.FRT.periph.id.enum
+Peripheral.INT.periph.id.enum
+Peripheral.OTH.periph.id.enum
+PeripheralBook
+PeripheralBookIdentification.book.code.enum
+PeripheralContent
+PeripheralDividedBook
+PeripheralDividedBookIdentification.book.code.enum
+PeripheralOther
+USX
+Verse
+VerseEnd
+cell.align.enum
+char.closed
+char.link
