+++ vitest.config.ts
@@ -1,0 +1,5 @@
+import { defineConfig } from 'vitest/config';
+
+export default defineConfig({
+    assetsInclude: ['**/*.rng'],
+});
